Fold DepthwiseConv into layers.py
Per review: a one-class file isn't warranted. Move DepthwiseConv (and its two
module-level helpers) into layers.py next to the other Strata building blocks,
delete depthwise_conv.py, and update the imports in strata.py and the test. Pure
relocation -- no code changes.
